Move useMap.tsx
